chore: move table view files
