+++ packages/compass-crud/src/components/document-table-view.less
@@ -1,3 +1,8 @@
+@import 'table-view/table-view-cell.less';
+@import 'table-view/table-view-cell-editor.less';
+@import 'table-view/table-view-row-actions.less';
+@import 'table-view/ag-grid.less';
+
 .document-table-view-container {
   width: 100%;
   height: 100%;
