+++ packages/compass-crud/src/components/index.less
@@ -3,7 +3,3 @@
 @import 'document.less';
 @import 'document-json-view.less';
 @import 'loading-indicator.less';
-@import 'table-view-cell.less';
-@import 'table-view-cell-editor.less';
-@import 'table-view-row-actions.less';
-@import 'ag-grid.less';
