+++ packages/compass-crud/src/components/table-view/table-view-cell.less
